+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.29.3] - 2018-11-07
+### Fixed
+- CXV1-15973 - Suggestions dropdown were not being closed on click outside of the component.
+- SuggestionsDropdown were being blocked by other panel when members table was empty.
+
 ## [0.29.2] - 2018-11-07
 ### Fixed
 - remove autocomplete suggestions onBlur